backend/requirements.txt
uvicorn>=0.30.6
fastapi>=0.115.0
sqlalchemy>=2.0.35
databases>=0.9.0
asyncpg>=0.29.0
alembic>=1.13.2
python-dotenv>=1.0.1
psycopg2-binary>=2.9.9
pydantic>=2.9.2
passlib>=1.7.4
bcrypt==4.0.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
httpx>=0.24.1
pyjwt==2.10.1
Jinja2==3.1.6
